feat(file-encryptor): include only minimal data in the second and subsequent chunks
The following data is not required when encrypting the second and subsequent chunks:

+ Multicodec-compliant identifier
+ Salt and options required to generate the key from the password.
+ Compression algorithm type

For this reason, only the minimum amount of data is stored in the header of the second and subsequent chunks.
With this improvement, the header length of the second and subsequent chunks was reduced by 60% (a simple test result showed that the 161-byte header was reduced to 64 bytes).