fix(file-encryptor): give up on creating a generic `mergeObjectWithoutUndefined()` function
I was unable to resolve a TypeScript compilation error caused by nested type definitions.
After spending a few hours trying different solutions, I realized that the function we needed was much simpler.
Now I have removed the `mergeObjectWithoutUndefined()` function and created a `normalizeOptions()` function in its place.
This should be sufficient for us, if only for the purpose of merging the input values with the default option values.