test(file-encryptor): skipping tests because we couldn't solve a bug in the `decryptStream()` function
The Transform stream emits an `error` event followed by a `close` event when an error occurs.
However, the `duplexify@4.1.1` package will emit a `finish` event before the `error` event.
Because of this, errors emitted from the decryptor are not received by the `stream.pipeline()` module method.
This problem can be solved by using the `pipe()` method or the `stream.pipeline()` module method to attach the Writable stream to the decryptor.
In most cases, the decryptor stream is connected to the Writable stream for use.
Therefore, it should not affect most users.
However, there may be other bugs that cause errors emitted by the decryptor to be ignored.
This bug has to be fixed.