chore(deps): use `@sounisi5011/run-if-supported` on npm
The `@sounisi5011/run-if-supported` must be built before it can be used.
This is why the `pnpm install` command will fail if it adds this of a submodule in the workspace to the dependencies of each package.
We can solve this problem by adding a `postinstall` script to `@sounisi5011/run-if-supported`, but the `postinstall` script is completely unnecessary when publishing to npm.
So, we will use the `@sounisi5011/run-if-supported` package published on npm instead.