chore(cli-utils-top-level-await): remove `@sounisi5011/run-if-supported` from the devDependencies
I noticed a circular reference in the project dependencies when I added `@sounisi5011/cli-utils-top-level-await` to the dependency of `@sounisi5011/run-if-supported` in https://github.com/sounisi5011/npm-packages/pull/130.
Due to this circular reference, the build using ultra-runner will fail.
ultra-runner has tried to build `@sounisi5011/run-if-supported` before `@sounisi5011/cli-utils-top-level-await`.
To solve this problem, remove `@sounisi5011/run-if-supported` from the `@sounisi5011/cli-utils-top-level-await` dependency.