test(jest-binary-data-matchers): always enable color output
ANSI colors are disabled on CI such as GitHub Actions.
As a result, tests that assume color output will fail.

To fix this problem, we tried to enable `chalk` by overriding the `FORCE_COLOR` environment variable in the test code.
https://github.com/chalk/supports-color#info

However, this fails due to Jest.
For this reason, we added the `--colors` option in the Jest CLI to always enable color output on CI.
https://jestjs.io/docs/cli#--colors