test(jest-binary-data-matchers): always enable color output (use `FORCE_COLOR` with `--colors` option)
ANSI colors are disabled on CI such as GitHub Actions.
As a result, tests that assume color output will fail.

To fix this problem, we added the `--colors` option in the Jest CLI.
https://jestjs.io/docs/cli#--colors

However, the color output was not enabled on GitHub Actions.
This is why we will also use the environment variable FORCE_COLOR` for `chalk` to try to enable color output.
https://github.com/chalk/supports-color#info