chore(typescript): disable new options in some packages
The following options are new in TypeScript 4.4 and require fixing some package contents:

+ `useUnknownInCatchVariables`
+ `exactOptionalPropertyTypes`

Since there are many things to fix (API changes, dependency updates, etc.), it should not be done only in this pull request.
Therefore, we decided to disable these options in each package and fix them in subsequent pull requests.