test(encrypted-archive): monitor memory usage of sub-processes instead of test code
When monitoring memory usage in the test code, the difference in memory usage may become negative due to garbage collection.
Also, Jest does not necessarily run each test in a different sub-process.

Therefore, we changed to run encryption and decryption in a sub-process and use the memory usage of the sub-process for testing.