revert: feat(run-if-supported): update source codes from CommonJS to ESM
A change in module type caused a large number of errors for which the cause could not be determined:

+ error TS7016: Could not find a declaration file for module 'npm-install-checks'.
+ error TS7016: Could not find a declaration file for module 'parse-json'.
+ error TS2339: Property 'object' does not exist on type 'typeof import(".../node_modules/ow/dist/index")'.

In particular, the cause of the error that the type definition of the parse-json package cannot be found is unknown. Because @types/parse-json exists as a dependency. In addition, the cross-spawn package has a similar structure, but no error occurs.

This reverts commit 8e1a66abfcb735113b4dc40b0bf18b7bed485101.