chore(npm-scripts): run ESLint twice when formatting
ESLint with the `--report-unused-disable-directives` option will remove unnecessary directive comments when formatting. However, indentations inserted for comments are not removed.
The removal of unneeded indentations is the formatting that dprint performs, but it is necessary to format again in ESLint after dprint's formatting.
Therefore, we run ESLint twice, once before and once after dprint, to try to format the source code properly.