test(encrypted-archive): skip older versions which do not support the current Node.js
Older `@sounisi5011/encrypted-archive` does not support some versions of Node.js. If a forward compatibility test is run on such a version of Node.js, the test will fail.
Therefore, the version of `@sounisi5011/encrypted-archive` that does not support the version of Node.js for which the test is being run should be skipped.