test(encrypted-archive): avoid forward compatibility test failures at release time (#462)
When we try to release a new version, it will generate test data for a version that does not yet exist on npm.
Forward compatibility testing is based on new test data and attempts to install a version of `@sounisi5011/encrypted-archive` that does not yet exist on npm.
This causes the forward compatibility test to fail only at release time.
To work around this, the latest version of `@sounisi5011/encrypted-archive` uses packages on local disk instead if installation from on npm fails.

Note: Ideally, we should use the "npm view" command to skip tests for packages that do not exist on npm. However, Jest is not yet capable of skipping tests asynchronously. see https://github.com/facebook/jest/issues/8604