chore(run-if-supported)!: support for older Node.js 14
The ow package in dependencies [supports Node.js 14.16 or later].
However, we have actually tried it and confirmed that it works with Node.js 14, which is smaller than this version.
So, change the minimum version of Node.js supported by `@sounisi5011/run-if-supported` to 14.13.1.
Node.js 14.13.1 is the minimum version that [supports `node:` imports].

BREAKING CHANGE: require Node.js `>=14.13.1`. This is the [smallest version that supports `node:` imports][supports `node:` imports].

[supports Node.js 14.16 or later]: https://github.com/sindresorhus/ow/blob/v1.0.0/package.json#L24-L26
[supports `node:` imports]: https://nodejs.org/api/esm.html#node-imports