refactor(@sounisi5011/encrypted-archive): define `CompressOptions` types in runtime dependent code
The `CompressOptions` type is strongly dependent on the Node.js "zlib" module.
If `CompressOptions` type is defined in the core code, the Node.js type definition cannot be eliminated from the core code.

Therefore, we changed the `CompressOptions` type to be defined in the runtime-dependent code (under `src/runtimes/`) instead of in the core code.
This should allow us to use different compression options type definitions for each platform, even if we support other platforms in the future.