ci(release): add fallback in case `outputs.tag_name` value from `google-github-actions/release-please-action@v3` is an empty string
We tried release for the first time since we updated `google-github-actions/release-please-action` to v3.
It failed to auto-release because the value of `outputs.tag_name` was an empty string.
https://github.com/sounisi5011/npm-packages/actions/runs/3818055716/jobs/6494701840#step:15:2

We don't know why this happens, so we added code to fall back to the next time the same thing happens.