refactor(@sounisi5011/encrypted-archive): should not use `import * as name from '...'` statement
When code using the `import * as name from '...'` statement ("namespace import" or commonly referred to as "import star") is transpiled with TypeScript, the build output will include the `__importStar` function.
So we should not use the `import * as name from '...'` statement as this will increase the build output file size.