refactor(@sounisi5011/encrypted-archive): rename `src/runtimes/browser/` to `src/runtimes/web-interoperable`
The directory name "browser" makes it seem as if the code is intended only for web browsers.
However, we will include in this directory code for any platform that conforms to WinterCG standards.
So we changed the directory name to "web-interoperable".

Note: I am not sure if the name "web-interoperable" is appropriate.
      This word is taken from the WinterCG's official name, "Web-interoperable Runtimes Community Group".
      But I'm not familiar with English, so this name could be wrong.
      "web-standard" might have been more appropriate.