build(@sounisi5011/encrypted-archive): run `protoc --version` command before parallel executing `protoc` commands
`@protobuf-ts/protoc` downloads the binary of the `protoc` command when the `protoc` command is first run after installation.
At this time, an ETXTBSY error may occur if the `protoc` command is executed in parallel.

To avoid this problem, run the `protoc --version` command before executing the `protoc` command in parallel.
This should complete the download of the `protoc` command.