build(@sounisi5011/ts-utils-is-property-accessible): fix the rollup.config.js file to support native Node ES modules
Rollup v3 loads configuration files as native ES modules.
This means that the default import of old CommonJS packages will not work as before.

By using the `--bundleConfigAsCjs` option, Rollup will load the configuration file as before.
However, now ESMs are becoming the standard.
So we decided to fix the configuration file.