fix(package.json): add "types" import condition to the `exports` field in `package.json`
If there is an `exports` field in `package.json`, TypeScript will now look for a type declaration file using the "types" import condition instead of the `types` field.
see https://www.typescriptlang.org/docs/handbook/esm-node.html#packagejson-exports-imports-and-self-referencing

Even if there is no "types" import condition, TypeScript will not use the `types` field.
So TypeScript **will not find a type declaration file**.

However, this problem does not happen with published packages.
so far, the published packages have the same name for a type declaration files and a `.js` files.
In this case, TypeScript will automatically find a type declaration file.

But if in future package updates we rename `.js` files and type declaration files, this problem will start to happen.
I don't know if we will remember this problem at that time.
We may release a package containing this problem without knowing it.

Therefore, add a "types" import condition to the `package.json` of all packages that require a type declaration file.