ci(update-license-year): update all `LICENSE` files
The `LICENSE` files do not necessarily exist only in the project root.
`LICENSE` files in subdirectories must also be automatically updated.

Some packages may use different license files.
For example, since it is currently 2023, a new package created this year must adopt a license starting in 2023.
In such cases, the `LICENSE` files are placed in the directory of each package.