chore(npm-scripts): use ultra-runner instead of Turborepo for linting and testing
Turborepo does not correctly display the output of the task on the console.
Turborepo's output disables ANSI escaping of the command being executed and does not display animations and/or colors.
In addition, the task log lines are longer due to the addition of ugly prefixes.

Turborepo's cache strategy is superior to Ultra Runner.
However, the output to the console is better with Ultra Runner.
In linting and testing, we prioritize console readability over execution speed and cache.