docs(@sounisi5011/encrypted-archive): rename `.edn` files to be converted by bytefield-svg to `.bytefield.edn`
edn (extensible data notation) is simply a data notation file.
Even if a file extension is `.edn`, it does not necessarily mean that the file is intended for conversion by bytefield-svg.
So we rename the file extension to `.bytefield.edn` to make it clear that the `.edn` file is for bytefield-svg.