chore(package.json): shouldn't use the `--yes` option with the `npx` command
The `npx` command included with npm v6 does not allow the `--yes` option.
However, the `npx` command included with npm v7 and later will print a prompt if the `--yes` option is not added.

To solve this problem, we decided to include `only-allow` in the dev dependencies for this project.
This will ensure that `only-allow` is executed correctly in each version of the `npx` command:

+ If using npm v6, the `--yes` option is not necessary.
  npx will run `only-allow` without prompting.
+ If using npm v7 or later, the `npm install` command will install the dependencies before running the `preinstall` script.
  `only-allow` is included in the dependencies, so npx will not fetch `only-allow` from remote.
  Therefore, npx will not print a prompt.