chore(package.json): instead of using the `--yes` option, input "y" into stdin of the `npx` command
The `npx` command included with npm v6 does not allow the `--yes` option.
However, the `npx` command included with npm v7 and later will print a prompt if the `--yes` option is not added.

To solve this problem, we decided to automatically input "y" into the stdin of the `npx` command.
This is the same solution as using the `yes` command in Unix.
However, the `yes` command is not available on Windows, so the `node` command is used to output "y" instead of it.

Note: The `npx` command included with npm v7 and later seems not to print a prompt just the Unix pipeline is connected.
      see https://github.com/npm/cli/blob/libnpmexec-v5.0.11/workspaces/libnpmexec/lib/index.js#L245-L262
      see https://github.com/npm/cli/blob/libnpmexec-v5.0.11/workspaces/libnpmexec/test/prompt.js#L167-L186
      Thus, without outputting "y", this problem can be solved with the following simple command: `node -e "" | npx only-allow pnpm`
      However, this behavior may not be the same for future `npx` commands.