ci(github actions): fix the name of steps for enable Corepack to be more understandable
The step name "Setup Corepack" indicates that it is setting up Corepack.
However, this step name does not clearly indicate that pnpm will be available automatically.
Contributors who are not familiar with Node.js will not be able to grasp what is possible by enabling Corepack.
So we fix the step names and clarify what we are enabling Corepack for.